--- tinyos/console.c
@@ -1,4 +1,0 @@
-void consoleinit(void)
-{
-    
-}
