--- tinyos/uart.h
@@ -1,6 +1,0 @@
-#ifndef UART_H
-# define UART_H
-
-void uart_send_char(char c);
-
-#endif
